+++ quantforge/__init__.py
@@ -727,6 +727,7 @@ from .reverse_ad import Var, reverse_gradient
 from .lbfgs import lbfgs
 from .reverse_jacobian import reverse_jacobian, reverse_hessian, reverse_gradient_vector
 from .newton_min import newton_min
+from .gauss_newton import gauss_newton
 from .dual import Dual, derivative as dual_derivative
 from .dual_calculus import dual_gradient, dual_newton
 from .hyperdual import HyperDual, derivatives as hyperdual_derivatives, second_derivative
@@ -1431,7 +1432,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "5.14.0"
+__version__ = "5.15.0"
 
 __all__ = [
     "OptionType",
@@ -2430,6 +2431,7 @@ __all__ = [
     "reverse_gradient_vector",
     "lbfgs",
     "newton_min",
+    "gauss_newton",
     "Dual",
     "dual_derivative",
     "dual_gradient",